Release 2.2.0

**Changelog**:
- Add auto-improve troops (works for both servers but TTWars is a better use case because of zero second upgrade). Read docs [here](https://tbs-docs.readthedocs.io/en/latest/improve.html)
- Redesign Village>Setting tab and add a checkbox to refresh setting to enable or disable
- Fix after first sleep, TBS wants to work 24/24
- Fix TBS use gold upgrade even with residence and similar building
- Fix TBS forgot to check the prerequisite building before build
- Fix TBS didn't know Travian forced to log out after a few inactive hours  

If you like my work, you can donate to me through [ko-fi.com/vinaghost](ko-fi.com/vinaghost)

+++ MainCore/Tasks/Update/UpdateInfo.cs
@@ -1,4 +1,5 @@
 ﻿using MainCore.Enums;
+using MainCore.Helper;
 using MainCore.Models.Database;
 using System;
 using System.Collections.Generic;
@@ -32,9 +33,15 @@ namespace MainCore.Tasks.Update
 
         public override void Execute()
         {
-            UpdateVillageList();
+            {
+                using var context = _contextFactory.CreateDbContext();
+                NavigateHelper.AfterClicking(_chromeBrowser, context, AccountId);
+            }
+            IsFail = true;
             UpdateAccountInfo();
+            UpdateVillageList();
             UpdateHeroInfo();
+            IsFail = false;
         }
 
         private void UpdateVillageList()
@@ -65,6 +72,7 @@ namespace MainCore.Tasks.Update
                 context.DeleteVillage(item.Id);
             }
 
+            var tribe = context.AccountsInfo.Find(AccountId).Tribe;
             foreach (var newVill in foundVills)
             {
                 context.Villages.Add(new Village()
@@ -76,6 +84,7 @@ namespace MainCore.Tasks.Update
                     Y = newVill.Y,
                 });
                 context.AddVillage(newVill.Id);
+                context.AddTroop(newVill.Id, tribe);
 
                 var tasks = _taskManager.GetList(AccountId).OfType<UpdateVillage>().ToList();
                 var task = tasks.FirstOrDefault(x => x.VillageId == newVill.Id);
@@ -147,10 +156,13 @@ namespace MainCore.Tasks.Update
             var adventures = context.Adventures.Count(x => x.AccountId == AccountId);
             if (numberAdventure == 0)
             {
-                var heroAdventures = context.Adventures.Where(x => x.AccountId == AccountId).ToList();
+                if (adventures != 0)
+                {
+                    var heroAdventures = context.Adventures.Where(x => x.AccountId == AccountId).ToList();
 
-                context.Adventures.RemoveRange(heroAdventures);
-                context.SaveChanges();
+                    context.Adventures.RemoveRange(heroAdventures);
+                    context.SaveChanges();
+                }
             }
             else if (adventures != numberAdventure)
             {
